fix: set policy field in analyze_ctx if one exists for repo
