+++ src/macaron/slsa_analyzer/checks/base_check.py
@@ -7,11 +7,6 @@ import logging
 from abc import abstractmethod
 from typing import Optional
 
-from sqlalchemy import Column, ForeignKey
-from sqlalchemy.orm import declarative_mixin, declared_attr
-from sqlalchemy.sql.sqltypes import Boolean, Integer, String
-
-from macaron.database.database_manager import ORMBase
 from macaron.slsa_analyzer.analyze_context import AnalyzeContext
 from macaron.slsa_analyzer.checks.check_result import CheckResult, CheckResultType, SkippedInfo, get_result_as_bool
 from macaron.slsa_analyzer.slsa_req import ReqName, get_requirements_dict
@@ -144,40 +139,3 @@ class BaseCheck:
             The result type of the check (e.g. PASSED).
         """
         raise NotImplementedError
-
-
-class CheckResultTable(ORMBase):
-    """Table to store the result of a check, is automatically added for each check."""
-
-    __tablename__ = "_check_result"
-    id = Column(Integer, primary_key=True, autoincrement=True)  # noqa: A003 # pylint: disable=invalid-name
-    check_id = Column(String, nullable=False)
-    repository = Column(Integer, ForeignKey("_repository.id"), nullable=False)
-    passed = Column(Boolean, nullable=False)
-    skipped = Column(Boolean, nullable=False)
-
-
-@declarative_mixin
-class CheckFactsTable:
-    """
-    Declarative mixin for check results.
-
-    All tables for check results must inherit this class, these fields are automatically filled in by the analyzer.
-    """
-
-    # pylint: disable=no-member
-
-    @declared_attr  # type: ignore
-    def id(self) -> Column:  # noqa: A003 # pylint: disable=invalid-name
-        """Check result id."""
-        return Column(Integer, primary_key=True, autoincrement=True)
-
-    @declared_attr  # type: ignore
-    def check_result(self) -> Column:
-        """Store the id of the repository to which the analysis pertains."""
-        return Column(Integer, ForeignKey("_check_result.id"), nullable=False)
-
-    @declared_attr  # type: ignore
-    def repository(self) -> Column:
-        """Store the id of the repository to which the analysis pertains."""
-        return Column(Integer, ForeignKey("_repository.id"), nullable=False)
