--- tests/policy_engine/resources/policies/policytest.cue
@@ -1,7 +1,0 @@
-{
-    predicate: {
-        builder: {
-            id: "https://github.com/slsa-framework/slsa-github-generator/.github/workflows/builder_go_slsa3.yml@refs/tags/v1.2.0"
-        }
-    }
-}
