+++ tests/policy_engine/resources/policies/valid/policytest.cue
@@ -1,0 +1,7 @@
+{
+    predicate: {
+        builder: {
+            id: "https://github.com/slsa-framework/slsa-github-generator/.github/workflows/builder_go_slsa3.yml@refs/tags/v1.2.2"
+        }
+    }
+}
